+++ src/onegov.town/onegov/town/models/__init__.py
@@ -4,7 +4,7 @@ from onegov.town.models.editor import Editor
 from onegov.town.models.file import File, FileCollection
 from onegov.town.models.form import BuiltinFormDefinition, CustomFormDefinition
 from onegov.town.models.image import Image, ImageCollection, Thumbnail
-from onegov.town.models.page import News, Topic
+from onegov.town.models.page import AtoZPages, News, Topic
 from onegov.town.models.page_move import PageMove
 from onegov.town.models.resource import DaypassResource
 from onegov.town.models.search import Search
@@ -15,6 +15,7 @@ from onegov.town.models.traitinfo import TraitInfo
 
 __all__ = [
     'AtoZ',
+    'AtoZPages',
     'BuiltinFormDefinition',
     'Clipboard',
     'CustomFormDefinition',