+++ src/onegov.town/onegov/town/models/page.py
@@ -2,6 +2,7 @@ from onegov.core import utils
 from onegov.page import Page
 from onegov.search import ORMSearchable
 from onegov.town.forms import LinkForm, PageForm
+from onegov.town.models.atoz import AtoZ
 from onegov.town.models.traitinfo import TraitInfo
 from onegov.town.models.extensions import (
     ContactExtension,
@@ -150,3 +151,24 @@ class News(Page, TraitInfo, SearchablePage,
         query = query.options(undefer('content'))
 
         return query
+
+
+class AtoZPages(AtoZ):
+
+    def get_title(self, item):
+        return item.title
+
+    def get_items(self):
+        from onegov.town.models.page import Topic
+
+        # XXX implement correct collation support on the database level
+        topics = self.request.app.session().query(Topic).all()
+        topics = sorted(topics, key=self.sortkey)
+
+        if self.request.is_logged_in:
+            return [topic for topic in topics if topic.trait == 'page']
+        else:
+            return [
+                topic for topic in topics if topic.trait == 'page'
+                and not topic.is_hidden_from_public
+            ]