+++ src/onegov.town/onegov/town/path.py
@@ -20,7 +20,7 @@ from onegov.libres.models import Resource
 from onegov.town.app import TownApp
 from onegov.town.converters import extended_date_converter
 from onegov.town.models import (
-    AtoZ,
+    AtoZPages,
     Clipboard,
     Editor,
     File,
@@ -250,6 +250,6 @@ def get_search(request, q='', page=0):
     return Search(request, q, page)
 
 
-@TownApp.path(model=AtoZ, path='/a-z')
+@TownApp.path(model=AtoZPages, path='/a-z')
 def get_a_to_z(request):
-    return AtoZ(request)
+    return AtoZPages(request)